Move curl_get() to the main MediaWikiBot class and change grabDeletedFiles to extend ExternalWikiGrabber.
