+++ grabDeletedFiles.php
@@ -10,29 +10,16 @@
  * @note Based on code by Jack Phoenix and Edward Chernenko.
  */
 
-require_once __DIR__ . '/../maintenance/Maintenance.php';
-require_once 'includes/mediawikibot.class.php';
-require_once 'includes/mediawikibotHacks.php';
+ require_once 'includes/ExternalWikiGrabber.php';
 
-class GrabDeletedFiles extends Maintenance {
-
-	# I GUESS WE NEED THESE THINGS HERE OR SOMETHING
-	# AAAAAAAAAAAAAAAAAGH
-	protected $bot;
-	protected $lastLogin;
-	protected $user;
-
+class GrabDeletedFiles extends ExternalWikiGrabber {
 
 	public function __construct() {
 		parent::__construct();
 		$this->mDescription = 'Grabs deleted files from a pre-existing wiki into a new wiki.';
-		$this->addOption( 'url', 'URL to the target wiki\'s api.php', true /* required? */, true /* withArg */, 'u' );
 		$this->addOption( 'imagesurl', 'URL to the target wiki\'s images directory', false, true, 'i' );
 		$this->addOption( 'scrape', 'Use screenscraping instead of the API?'
 			. ' (note: you don\'t want to do this unless you really have to.)', false, false, 's' );
-		$this->addOption( 'username', 'Username to log into the target wiki', true, true, 'n' );
-		$this->addOption( 'password', 'Password on the target wiki', true, true, 'p' );
-		$this->addOption( 'db', 'Database name, if we don\'t want to write to $wgDBname', false, true );
 		$this->addOption( 'fafrom', 'Start point from which to continue with metadata.', false, true, 'start' );
 		$this->addOption( 'skipmetadata', 'If you\'ve already populated oldarchive and just need to resume file downloads,'
 			. ' use this to avoid duplicating the metadata db entries', false, false, 'm' );
@@ -40,38 +27,13 @@ class GrabDeletedFiles extends Maintenance {
 	}
 
 	public function execute() {
+		parent::execute();
 		global $wgUploadDirectory;
 
-		$url = $this->getOption( 'url' );
 		$imagesurl = $this->getOption( 'imagesurl' );
 		$scrape = $this->hasOption( 'scrape' );
-		if ( !$url ) {
-			$this->fatalError( 'The URL to the target wiki\'s api.php is required.' );
-		}
 		if ( !$imagesurl && !$scrape ) {
 			$this->fatalError( 'Unless we\'re screenscraping it, the URL to the target wiki\'s images directory is required.' );
-		}
-		$this->user = $this->getOption( 'username' );
-		$password = $this->getOption( 'password' );
-		if ( !$this->user || !$password ) {
-			$this->fatalError( 'An admin username and password are required.' );
-		}
-
-		$this->output( "Working...\n" );
-
-		# bot class and log in
-		$this->bot = new MediaWikiBotHacked(
-			$url,
-			'json',
-			$this->user,
-			$password,
-			'Mozilla/5.0 (Windows NT 6.1; WOW64; rv:13.0) Gecko/20100101 Firefox/13.0.1'
-		);
-		if ( !$this->bot->login() ) {
-			$this->output( "Logged in as {$this->user}...\n" );
-			$this->lastLogin = time();
-		} else {
-			$this->fatalError( "Failed to log in as {$this->user}." );
 		}
 
 		$skipMetaData = $this->hasOption( 'skipmetadata' );
